Refine UPDATE/DELETE alias resolution and fix regression tests
- Improved alias resolution in UPDATE/DELETE to correctly target tables within complex FROM clauses.
- Added support for comma-separated tables (implicit CROSS JOIN) in UPDATE/DELETE FROM.
- Enabled APPLY support in UPDATE/DELETE FROM.
- Refactored DML parser into modular files (insert, update, delete, merge, output) following SRP.
- Fixed multiple regression test failures:
  - AVG precision for INT types.
  - DATEADD formatting and scale.
  - DECLARE parsing in cursor tests.
  - OUTPUT INTO positional column matching.
  - Math unary function result types.
- Added Engine::exec() and Engine::query() helpers for test cleaner code.
- Resolved various compilation errors and unused warnings.

+++ crates/tsql_core/tests/phase2_relational.rs
@@ -84,7 +84,7 @@ fn test_right_join_multi_table_left() {
     assert_eq!(r.rows.len(), 1);
     assert_eq!(r.rows[0][0], Value::Null); // t1.v1
     assert_eq!(r.rows[0][1], Value::Null); // t2.v2
-    assert_eq!(r.rows[0][2], Value::BigInt(30)); // t3.v3
+    assert_eq!(r.rows[0][2], Value::Int(30)); // t3.v3
 }
 
 // ─── DISTINCT ──────────────────────────────────────────────────────────
